get_details {YoutubeLitev3}	R Documentation

Get detail information about the Youtube Video

Description

The functions will provide the following information about a Video : Id, Publish Date, Channel ID, Title of the Video, Channel Title, and Category Id of the video.

Usage

get_details(video_url = Youtube_Video_URL)

Arguments

video_url
URL of the Youtube Video.

Note

Users need to run set_key function to access to Youtube database.

Author(s)

Tarndeep Chahal, Vincent Phan, Shahid Jahagirdar

Examples

set_key('AIzaSyAdy65AkdC-E8--0DYB5osklXQFQOKACVs')
get_details('https://www.youtube.com/watch?v=d-eUoZzFutc')

[Package YoutubeLitev3 version 0.1.0 Index]