set_key {YoutubeLitev3}	R Documentation

Setup Youtube API Key

Description

This function uses Youtube API Key to set up the access to Youtube Database. Users need to register a API Key before using the function.

Usage

set_key(key=API_key)

Arguments

key
Youtube API key, an user need to register for an API key from Youtube website

Details

This function will set up the authorization process to access to Youtube database by using the Youtube API provided from the user. This function need to be implemented before running any function from this package

Author(s)

Tarndeep Chahal, Vincent Phan, Shahid Jahagirdar.

Examples

set_key('AIzaSyAdy65AkdC-E8--0DYB5osklXQFQOKACVs')

[Package YoutubeLitev3 version 0.1.0 Index]